pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nika-ai"
version = "0.1.0"
description = "Nika — Lightweight fully local autonomous AI agent"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.3.0",
    "textual>=0.61.0",
    "rich>=13.7.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "chromadb>=0.5.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "trafilatura>=1.9.0",
    "duckduckgo-search>=6.1.0",
    "psutil>=5.9.8",
    "pyperclip>=1.8.2",
    "weasyprint>=62.0",
    "jinja2>=3.1.4",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "typer>=0.12.3",
    "loguru>=0.7.2",
    "tiktoken>=0.7.0",
    "edge-tts>=6.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
]

[project.scripts]
nika = "nika.main:app"

[tool.hatch.build.targets.wheel]
packages = ["nika"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
